+++ src/pages/vendorDashboard/vendorInventory.tsx
@@ -1,3 +1,6 @@
+import { useState, useEffect } from "react"
+import axios from 'axios';
+
 import Header from "../vendorcomponents/Header"
 import { PencilLine, Trash2, ListFilter } from "lucide-react"
 
@@ -10,12 +13,84 @@ import {
   TableRow,
 } from "../../../src/components/ui/table"
 
+interface Product {
+  id: number;
+  categoryId: number;
+  userId: number;
+  name: string;
+  description: string;
+  price: string;
+  stock: number;
+  image: string;
+  createdAt: string;
+  updatedAt: string;
+  category: {
+    id: number;
+    categoryName: string;
+    description: string;
+  };
+  user: {
+    id: number;
+    firstName: string;
+    lastName: string;
+    username: string;
+    email: string;
+    password: string;
+    userType: string;
+    registerType: string;
+    socialId: null | string;
+    registrationDate: string;
+    phoneNumber: string;
+    confirmPassword: null | string;
+    isVerified: string;
+    verificationToken: string;
+    verificationTokenExpiresAt: string;
+    resetToken: null | string;
+    resetTokenExpiresAt: null | string;
+  };
+}
+
+interface ApiResponse {
+  status: string;
+  code: number;
+  data: Product[];
+}
+
 const VendorInventory = () => {
+
+    const [products, setProducts] = useState<Product[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await axios.get<ApiResponse>('https://renergy-hub-express-backend.onrender.com/api/v1/products'); // Replace with your actual API endpoint
+        setProducts(response.data.data); 
+      } catch (err) {
+        setError('Failed to fetch products.');
+        console.error(err);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+   if (isLoading) {
+    return <div>Loading...</div>;
+  }
+
+  if (error) {
+    return <div>Error: {error}</div>;
+  }
+
   return (
     <div className="flex-1 overflow-auto relative z-10">
       <Header title='Inventory' />
 
-      <div className="bg-gray-200 h-full mx-auto my-auto pb-2">
+      <div className="bg-gray-200 h-auto  mx-auto my-auto pb-4">
         <div className="pt-8 w-[98%] flex justify-between flex-wrap mx-auto my-auto">
             <section className="bg-white rounded-md w-[98%] h-auto mx-auto">
                 <h2 className="text-xl ml-2 mt-2">
@@ -26,6 +101,7 @@ const VendorInventory = () => {
                         <h3 className="text-xl ml-2 mt-2 text-green-800">
                             Categories
                         </h3>
+                        
                         <Table className="bg-white rounded-lg mt-1 mb-0">
                             <TableHeader>
                                 <TableRow className="text-md">
@@ -120,29 +196,28 @@ const VendorInventory = () => {
                         </button>
                     </div>
                 </div>
-            
-            <Table>
+            {products.map((product) => (
+            <Table  key={product.id}>
                 <TableHeader>
                     <TableRow className="text-lg">
                     <TableHead className="text-black">Product</TableHead>
                     <TableHead className="text-black">Buying Price</TableHead>
                     <TableHead className="text-black">Quantity</TableHead>
-                    <TableHead className="text-black">Treshold Value</TableHead>
                     <TableHead className="text-black">Availability</TableHead>
                     </TableRow>
                 </TableHeader>
                 <TableBody>
                     <TableRow>
-                    <TableCell className="text-black font-medium border-b-2 pb-1">Fireman 380W Solar Panel</TableCell>
-                    <TableCell className="text-black font-medium border-b-2 pb-1">NGN250K</TableCell>
-                    <TableCell className="text-black font-medium border-b-2 pb-1">43 Packets</TableCell>
-                    <TableCell className="text-black font-medium border-b-2 pb-1">12 Packets</TableCell>
-                    <TableCell className="font-medium border-b-2 pb-1 text-green-600">In Stock</TableCell>
+                    <TableCell className="text-black font-medium border-b-2 pb-1 w-1/5">{product.name}</TableCell>
+                    <TableCell className="text-black font-medium border-b-2 pb-1 w-1/5">NGN{product.price}</TableCell>
+                    <TableCell className="text-black font-medium border-b-2 pb-1 w-1/5">{product.stock} Packets</TableCell>
+                    <TableCell className="font-medium border-b-2 pb-1 w-1/5">{product.stock > 0 ? <span className="text-green-600">In Stock</span> : <span className="text-red-600">Out Of Stock</span>}</TableCell>
                     <TableCell className="text-green-800 font-medium border-b-2 pb-1 w-8"><PencilLine /></TableCell>
                     <TableCell className="text-red-800 font-medium border-b-2 pb-1 w-8"><Trash2 /></TableCell>
                     </TableRow>
                 </TableBody>
             </Table>
+             ))}
         </div>
 
         </div>
